refactor: move credential validation and ProviderType to shared crates
Move the last provider-blocking types to tensorzero-inference-types:

- skip_credential_validation / with_skip_credential_validation /
  e2e_skip_credential_validation → credential_validation module
- ProviderType enum → credentials module

Update core tests to use with_skip_credential_validation() instead of
directly accessing the SKIP_CREDENTIAL_VALIDATION task-local.
