+++ crates/tensorzero-core/src/config/mod.rs
@@ -91,29 +91,9 @@ pub mod unwritten;
 
 pub use namespace::Namespace;
 
-tokio::task_local! {
-    /// When set, we skip performing credential validation in model providers
-    /// This is used when running in e2e test mode, and by the 'evaluations' binary
-    /// We need to access this from async code (e.g. when looking up GCP SDK credentials),
-    /// so this needs to be a tokio task-local (as a task may be moved between threads)
-    ///
-    /// Since this needs to be accessed from a `Deserialize` impl, it needs to
-    /// be stored in a `static`, since we cannot pass in extra parameters when calling `Deserialize::deserialize`
-    static SKIP_CREDENTIAL_VALIDATION: ();
-}
-
-pub fn skip_credential_validation() -> bool {
-    // tokio::task_local doesn't have an 'is_set' method, so we call 'try_with'
-    // (which returns an `Err` if the task-local is not set)
-    SKIP_CREDENTIAL_VALIDATION.try_with(|()| ()).is_ok()
-}
-
-/// Runs the provider future with credential validation disabled
-/// This is safe to repeatedly nest (e.g. `with_skip_credential_validation(async move { with_skip_credential_validation(f).await })`),
-/// the original credential validation behavior will be restored after the outermost future completes
-pub async fn with_skip_credential_validation<T>(f: impl Future<Output = T>) -> T {
-    SKIP_CREDENTIAL_VALIDATION.scope((), f).await
-}
+pub use tensorzero_inference_types::credential_validation::{
+    skip_credential_validation, with_skip_credential_validation,
+};
 
 /// Configuration for the autopilot system.
 #[derive(Clone, Debug, Default, Deserialize, PartialEq, Serialize)]
@@ -1354,14 +1334,7 @@ async fn process_uninitialized_config(
     })
 }
 
-/// In e2e test mode, we skip credential validation by default.
-/// This can be overridden by setting the `TENSORZERO_E2E_CREDENTIAL_VALIDATION` environment variable to `1`.
-/// Outside of e2e test mode, we leave the behavior unchanged (other parts of the codebase might still
-/// skip credential validation, e.g. when running in relay mode).
-pub fn e2e_skip_credential_validation() -> bool {
-    cfg!(any(test, feature = "e2e_tests"))
-        && !std::env::var("TENSORZERO_E2E_CREDENTIAL_VALIDATION").is_ok_and(|x| x == "1")
-}
+pub use tensorzero_inference_types::credential_validation::e2e_skip_credential_validation;
 
 impl Config {
     /// Constructs a new `Config`, as if from an empty config file.
